Updates progress for changes in automakemkv
Updates to automakemkv to use makemkvcon progress information required
renaming various signals and methods in the ProgressDialog classes
of cdripper and automakemkv so that they would not clash in a combined
class.

Updates so the the ProgressDialog of this package is an explit class
as having multiple inheritance with pyqt objects does not work.