Updates for new filegen, improve settings
Updates to autoMakeMKV have enabled file output convention setting
for either video_utils or plex. Have updated the keywords to the
main program to implement this. The 'filegen' keyword has been
removed and now the 'convention' (str) keyword is used to set either
'video_utils' or 'plex' convention. This can be changed in the Settings
dialog.

Have improved the settings dialog by moving the cdRipper and autoMakeMKV
settings (audio and video, respectively) to tabs in the settings dialog.
This was possible due to changes in the API of those respective packages.