Updates for multiple OS support
Adds codes for supporting both linux and windows machines.
This includes moving to more event, slot/signal based program
flow.

There have been updates to the cdripper and automakemkv APIs, namely
in the progress widgets, that had to be updated here.

The eject function on linux now uses Popen to call 'eject' so that
the GUI does not lock up.

Adds app and tray icons to the program