Updates for changes to user settings
autoMakeMKV had a major overhaul of the user settings management,
moving to a package-wide class to manage the settings. cdRipper has
also been moved to this paradigm.

Updates here take full advantage of those changes, with the only
caveat being that there is no longer a combied (autoripper) settings
file and instead the cdripper and automakemkv settings are updated
when changes are made in autoripper.