Fixes bugs in fail/success methods
When signaling video fail/success, was missing arguments to
the slots/methods that had been added so can display the file name
for success/failure. This has been fixed.

There is a new issue that must be check in autoMakeMKV where cancel
of process from progress bar cause QThread to be killed out of
turn and crash program.